+++ classes/local/participant/entities/student.php
@@ -478,7 +478,6 @@ class student extends participant {
         return $this->nextlesson;
     }
 
-
     /**
      * Return student's last lesson module completed.
      *
@@ -486,7 +485,7 @@ class student extends participant {
      */
     public function get_last_lesson_mod_complete() {
         $lessonmodscompleted = $this->vault->get_student_completed_lesson_ids($this->courseid, $this->userid);
-        if (empty($lessonmodscompleted)) {
+        if (empty($lessonmodscompleted) || empty($lessonmodscompleted->lessonmods)) {
             return get_string('nolessonrecord', 'local_booking');
         }
         $completedlessonmods = explode(',', $lessonmodscompleted->lessonmods);