+++ classes/local/session/data_access/booking_vault.php
@@ -187,9 +187,9 @@ class booking_vault implements booking_vault_interface {
         if (!empty($booking->get_exercise_id())) {
             $conditions['exerciseid'] = $booking->get_exercise_id();
         }
-        if (!empty($booking->active())) {
-            $conditions['active'] = '1';
-        }
+        // if (!empty($booking->active())) {
+        //     $conditions['active'] = '1';
+        // }
 
         return $DB->get_record(static::DB_BOOKINGS, $conditions);
     }